@@ -10,37 +10,39 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^16.2.0",
-    "@angular/common": "^16.2.0",
-    "@angular/compiler": "^16.2.0",
-    "@angular/core": "^16.2.0",
-    "@angular/forms": "^16.2.0",
-    "@angular/platform-browser": "^16.2.0",
-    "@angular/platform-browser-dynamic": "^16.2.0",
-    "@angular/router": "^16.2.0",
+    "@angular-eslint/schematics": "^17.2.1",
+    "@angular/animations": "^17.1.0",
+    "@angular/common": "^17.1.0",
+    "@angular/compiler": "^17.1.0",
+    "@angular/core": "^17.1.0",
+    "@angular/forms": "^17.1.0",
+    "@angular/platform-browser": "^17.1.0",
+    "@angular/platform-browser-dynamic": "^17.1.0",
+    "@angular/router": "^17.1.0",
     "@fortawesome/fontawesome-free": "^6.5.1",
-    "@ng-bootstrap/ng-bootstrap": "^15.1.2",
-    "@okta/okta-angular": "^5.2.0",
+    "@ng-bootstrap/ng-bootstrap": "^16.0.0",
+    "@okta/okta-angular": "^6.3.2",
     "@okta/okta-auth-js": "^7.5.0",
-    "@okta/okta-signin-widget": "^6.2.0",
+    "@okta/okta-signin-widget": "^7.14.0",
     "@popperjs/core": "^2.11.6",
+    "angular-font-awesome": "^3.1.2",
     "bootstrap": "^5.3.2",
+    "font-awesome": "^4.7.0",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.13.0"
+    "zone.js": "~0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.1.0",
-    "@angular/cli": "^16.2.6",
-    "@angular/compiler-cli": "^16.2.0",
-    "@angular/localize": "^16.2.12",
-    "@types/jasmine": "~4.3.0",
-    "jasmine-core": "~4.6.0",
+    "@angular-devkit/build-angular": "^17.1.1",
+    "@angular/cli": "^17.1.1",
+    "@angular/compiler-cli": "^17.1.0",
+    "@types/jasmine": "~5.1.0",
+    "jasmine-core": "~5.1.0",
     "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "typescript": "~5.1.3"
+    "typescript": "~5.3.2"
   }
 }
