@@ -1,15 +1,16 @@
-import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
+import { Injector, NgModule } from '@angular/core';
 
 import { AppComponent } from './app.component';
 import { ProductListComponent } from './components/product-list/product-list.component';
-import { HttpClientModule } from '@angular/common/http'
+import { HttpClientModule } from '@angular/common/http';
 import { ProductService } from './services/product.service';
 
 import { Routes, RouterModule } from '@angular/router';
 import { ProductCategoryMenuComponent } from './components/product-category-menu/product-category-menu.component';
 import { SearchComponent } from './components/search/search.component';
 import { ProductDetailsComponent } from './components/product-details/product-details.component';
+
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { CartStatusComponent } from './components/cart-status/cart-status.component';
 import { CartDetailsComponent } from './components/cart-details/cart-details.component';
@@ -18,26 +19,46 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { LoginComponent } from './components/login/login.component';
 import { LoginStatusComponent } from './components/login-status/login-status.component';
 
-import { OktaAuthModule, OktaCallbackComponent, OKTA_CONFIG } from '@okta/okta-angular';
+import {
+  OktaAuthModule,
+  OktaCallbackComponent,
+  OKTA_CONFIG,
+  OktaAuthGuard
+} from '@okta/okta-angular';
+
 import { OktaAuth } from '@okta/okta-auth-js';
+
 import myAppConfig from './config/my-app-config';
+import { MembersPageComponent } from './components/members-page/members-page.component';
+import { Router } from '@angular/router';
 
 const oktaConfig = myAppConfig.oidc;
+
 const oktaAuth = new OktaAuth(oktaConfig);
 
-const routes: Routes = [
-   {path: 'login/callback', component: OktaCallbackComponent},
-   {path: 'login', component: LoginComponent},
+function sendToLoginPage(oktaAuth: OktaAuth, injector: Injector){
+  //use injector to access any service available within you app
+  const router = injector.get(Router);
 
-   {path: 'checkout', component: CheckoutComponent},
-   {path: 'cart-details', component: CartDetailsComponent},
-   {path: 'products/:id', component: ProductDetailsComponent},
-   {path: 'search/:keyword', component: ProductListComponent},
-   {path: 'category/:id', component: ProductListComponent},
-   {path: 'category', component: ProductListComponent},
-   {path: 'products', component: ProductListComponent},
-   {path: '', redirectTo: '/products', pathMatch: 'full'},
-   {path: '**', redirectTo: '/products', pathMatch: 'full'}
+  //Redirect the user to your custom login page
+  router.navigate(['/login']);
+}
+
+const routes: Routes = [
+
+  {path: 'login/callback', component: OktaCallbackComponent},
+  {path: 'login', component: LoginComponent},
+  {path: 'members', component: MembersPageComponent, canActivate: [OktaAuthGuard], data: {onAuthRequired: sendToLoginPage}},
+
+  { path: 'checkout', component: CheckoutComponent },
+  { path: 'cart-details', component: CartDetailsComponent },
+  { path: 'products/:id', component: ProductDetailsComponent },
+  { path: 'search/:keyword', component: ProductListComponent },
+  { path: 'category/:id', component: ProductListComponent },
+  { path: 'category', component: ProductListComponent },
+  { path: 'products', component: ProductListComponent },
+  { path: '', redirectTo: '/products', pathMatch: 'full' },
+  { path: '**', redirectTo: '/products', pathMatch: 'full' }
 ];
 
 @NgModule({
@@ -59,8 +80,9 @@ const routes: Routes = [
     HttpClientModule,
     NgbModule,
     ReactiveFormsModule,
+    OktaAuthModule.forRoot({oktaAuth})
   ],
-  providers: [ProductService, { provide: OKTA_CONFIG, useValue: { oktaAuth } }],
+  providers: [ProductService, { provide: OKTA_CONFIG, useValue: { oktaAuth }}],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
