@@ -1,9 +1,8 @@
-import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
-import { of } from 'rxjs';
+import { Observable, of } from 'rxjs';
+import { HttpClient } from '@angular/common/http';
 import { Country } from '../common/country';
+import { map } from 'rxjs/operators';
 import { State } from '../common/state';
 
 @Injectable({
@@ -16,8 +15,9 @@ export class Luv2ShopFormService {
 
   constructor(private httpClient: HttpClient) { }
 
-  getCountries(): Observable<Country[]>{
-    return this.httpClient.get<GetResponseContries>(this.countriesUrl).pipe(
+  getCountries(): Observable<Country[]> {
+
+    return this.httpClient.get<GetResponseCountries>(this.countriesUrl).pipe(
       map(response => response._embedded.countries)
     );
   }
@@ -32,36 +32,47 @@ export class Luv2ShopFormService {
     );
   }
 
-  getCreditCardMonths(startMonth: number): Observable<number[]>{
-    let data: number[] = [];
 
-    //build an array for Month dropdown list
-    for(let theMonth = startMonth; theMonth <= 12; theMonth++){
+  getCreditCardMonths(startMonth: number): Observable<number[]> {
+
+    let data: number[] = [];
+    
+    // build an array for "Month" dropdown list
+    // - start at current month and loop until 
+
+    for (let theMonth = startMonth; theMonth <= 12; theMonth++) {
       data.push(theMonth);
     }
-    //the "of" operator from rxjs, will wrap an object as an Observable
+
     return of(data);
   }
 
-  getCreditCardYears(): Observable<number[]>{
+  getCreditCardYears(): Observable<number[]> {
+
     let data: number[] = [];
 
-    //build an array for 'year' dropdown list
+    // build an array for "Year" downlist list
+    // - start at current year and loop for next 10 years
+    
     const startYear: number = new Date().getFullYear();
     const endYear: number = startYear + 10;
 
-    for(let theYear = startYear; theYear <= endYear; theYear++){
+    for (let theYear = startYear; theYear <= endYear; theYear++) {
       data.push(theYear);
     }
+
     return of(data);
   }
+
 }
-interface GetResponseContries{
+
+interface GetResponseCountries {
   _embedded: {
     countries: Country[];
   }
 }
-interface GetResponseStates{
+
+interface GetResponseStates {
   _embedded: {
     states: State[];
   }
