@@ -1,16 +1,7 @@
-import { CartItem } from "./cart-item";
-
 export class OrderItem {
 
-    imageUrl!: string;
-    unitPrice!: number;
-    quantity!: number;
-    productId!: number;
-
-    constructor(cartItem: CartItem){
-        this.imageUrl = cartItem.imageUrl;
-        this.unitPrice = cartItem.unitPrice;
-        this.quantity = cartItem.quantity;
-        this.productId = cartItem.id;
-    }
+    constructor(public imageUrl: string,
+        public unitPrice: number,
+        public quantity: number,
+        public productId: string) { }
 }
