@@ -1,5 +1,5 @@
 export class Order {
 
-    totalQuantity!: number;
-    totalPrice!: number;
+    constructor (public totalQuantity: number,
+        public totalPrice: number) { }
 }
