@@ -1,4 +1,7 @@
 export class ProductCategory {
 
-    constructor(public id: number, public categoryName: string){}
+    constructor(public id: number,
+        public categoryName: string) {
+
+}
 }
