@@ -1,14 +1,15 @@
 export class Product {
 
-    constructor(
-        public id: number,
-        public sku: string,
-        public name: string,
-        public description: string,
-        public unitPrice: number,
-        public imageUrl: string,
-        public active: boolean,
-        public unitsInStock: number,
-        public dateCreated: Date,
-        public lastUpdated: Date) { }
+    constructor(public id?: string,
+        public sku?: string,
+        public name?: string,
+        public description?: string,
+        public unitPrice?: number,
+        public imageUrl?: string,
+        public active?: boolean,
+        public unitsInStock?: number,
+        public dateCreated?: Date,
+        public lastUpdated?: Date
+) {
+}
 }
