@@ -1,15 +1,12 @@
-import { Address } from "./address";
-import { Customer } from "./customer";
-import { Order } from "./order";
-import { OrderItem } from "./order-item";
-
+import { Address } from './address';
+import { Customer } from './customer';
+import { Order } from './order';
+import { OrderItem } from './order-item';
 
 export class Purchase {
-
     customer!: Customer;
     shippingAddress!: Address;
     billingAddress!: Address;
     order!: Order;
-    orderItems!: OrderItem[];
-
+    orderItems!: OrderItem[]; 
 }
