@@ -1,5 +1,4 @@
 export class State {
-
-    id!: number;
-    name!: string;
+    constructor(public id: number,
+        public name: string) { }
 }
