@@ -1,53 +1,52 @@
 import { Component, OnInit } from '@angular/core';
 import { CartItem } from 'src/app/common/cart-item';
-import { CartService } from 'src/app/services/cart-service';
+import { CartService } from 'src/app/services/cart.service';
 
 @Component({
   selector: 'app-cart-details',
   templateUrl: './cart-details.component.html',
   styleUrls: ['./cart-details.component.css']
 })
-export class CartDetailsComponent implements OnInit{
-  
+export class CartDetailsComponent implements OnInit {
+
   cartItems: CartItem[] = [];
   totalPrice: number = 0;
   totalQuantity: number = 0;
 
-  constructor(private cartService: CartService){}
-  
-  
+  constructor(private cartService: CartService) { }
+
   ngOnInit(): void {
     this.listCartDetails();
   }
 
   listCartDetails() {
-    //get a handle to the cart items
+
+    // get a handle to the cart items
     this.cartItems = this.cartService.cartItems;
 
-    //subscribe to the cart totalPrice
+    // subscribe to the cart totalPrice
     this.cartService.totalPrice.subscribe(
       data => this.totalPrice = data
     );
 
-    //subscribe to the cart total quantity
-    this.cartService.totalQuantity.subscribe(
+    // subscribe to the cart totalQuantity
+    this.cartService.totalQuantity.subscribe( 
       data => this.totalQuantity = data
     );
 
-    //compute cart total price and total quantity
+    // compute cart total price and quantity
     this.cartService.computeCartTotals();
   }
 
-  incrementQuantity(theCartItem: CartItem){
+  incrementQuantity(theCartItem: CartItem) {
     this.cartService.addToCart(theCartItem);
   }
 
-  decrementQuantity(theCartItem: CartItem){
+  decrementQuantity(theCartItem: CartItem) {
     this.cartService.decrementQuantity(theCartItem);
   }
 
   remove(theCartItem: CartItem) {
-    this.cartService.remove(theCartItem);  
+    this.cartService.remove(theCartItem);
   }
-
 }
