@@ -1,4 +1,4 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { CartStatusComponent } from './cart-status.component';
 
@@ -6,10 +6,14 @@ describe('CartStatusComponent', () => {
   let component: CartStatusComponent;
   let fixture: ComponentFixture<CartStatusComponent>;
 
-  beforeEach(() => {
+  beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [CartStatusComponent]
-    });
+      declarations: [ CartStatusComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
     fixture = TestBed.createComponent(CartStatusComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
