@@ -1,35 +1,34 @@
 import { Component, OnInit } from '@angular/core';
-import { CartService } from 'src/app/services/cart-service';
+import { CartService } from 'src/app/services/cart.service';
 
 @Component({
   selector: 'app-cart-status',
   templateUrl: './cart-status.component.html',
   styleUrls: ['./cart-status.component.css']
 })
-export class CartStatusComponent implements OnInit{
+export class CartStatusComponent implements OnInit {
 
   totalPrice: number = 0.00;
   totalQuantity: number = 0;
 
-  constructor(private cartService: CartService){}
-  
+  constructor(private cartService: CartService) { }
+
   ngOnInit(): void {
     this.updateCartStatus();
   }
 
-  updateCartStatus(){
-    //subscribe to the cart status totalPrice
+  updateCartStatus() {
 
+    // subscribe to the cart totalPrice
     this.cartService.totalPrice.subscribe(
       data => this.totalPrice = data
     );
 
-    //subscribe to the cart status totalQuantity
+    // subscribe to the cart totalQuantity
     this.cartService.totalQuantity.subscribe(
       data => this.totalQuantity = data
     );
-  }
 
-  
+  }
 
 }
