@@ -1,15 +1,15 @@
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
-import { Router } from '@angular/router';
+import { FormGroup, FormBuilder, FormControl, Validators } from '@angular/forms';
+import { Luv2ShopFormService } from 'src/app/services/luv2-shop-form.service';
 import { Country } from 'src/app/common/country';
+import { State } from 'src/app/common/state';
+import { Luv2ShopValidators } from 'src/app/validators/luv2-shop-validators';
+import { CartService } from 'src/app/services/cart.service';
+import { CheckoutService } from 'src/app/services/checkout.service';
+import { Router } from '@angular/router';
 import { Order } from 'src/app/common/order';
 import { OrderItem } from 'src/app/common/order-item';
 import { Purchase } from 'src/app/common/purchase';
-import { State } from 'src/app/common/state';
-import { CartService } from 'src/app/services/cart-service';
-import { CheckoutService } from 'src/app/services/checkout.service';
-import { Luv2ShopFormService } from 'src/app/services/luv2-shop-form.service';
-import { Luv2ShopValidators } from 'src/app/validators/luv2-shop-validators';
 
 @Component({
   selector: 'app-checkout',
@@ -19,61 +19,77 @@ import { Luv2ShopValidators } from 'src/app/validators/luv2-shop-validators';
 export class CheckoutComponent implements OnInit {
 
   checkoutFormGroup!: FormGroup;
-  shippingAddressStates: State[] = [];
-  billingAddressStates: State[] = [];
 
   totalPrice: number = 0;
   totalQuantity: number = 0;
-
+  
   creditCardYears: number[] = [];
   creditCardMonths: number[] = [];
 
   countries: Country[] = [];
-  states: State[] = [];
 
-
-  constructor(private formBuilder: FormBuilder, 
-              private luv2ShopFormService: Luv2ShopFormService, 
+  shippingAddressStates: State[] = [];
+  billingAddressStates: State[] = [];
+    
+  constructor(private formBuilder: FormBuilder,
+              private luv2ShopFormService: Luv2ShopFormService,
               private cartService: CartService,
               private checkoutService: CheckoutService,
-              private router: Router) {}
+              private router: Router) {
+               }
 
   ngOnInit(): void {
-
+    
     this.reviewCartDetails();
 
     this.checkoutFormGroup = this.formBuilder.group({
       customer: this.formBuilder.group({
-        firstName: new FormControl('', [Validators.required, Validators.minLength(2), Luv2ShopValidators.notOnlyWhitespace]),
-        lastName: new FormControl('', [Validators.required, Validators.minLength(2), Luv2ShopValidators.notOnlyWhitespace]),
-        email: new FormControl('', [Validators.required, Validators.pattern('^[a-z0-9._%+-]+@[a-z0-9.-]+\\.[a-z]{2,4}$')])
+        firstName: new FormControl('', 
+                              [Validators.required, 
+                               Validators.minLength(2), 
+                               Luv2ShopValidators.notOnlyWhitespace]),
+
+        lastName:  new FormControl('', 
+                              [Validators.required, 
+                               Validators.minLength(2), 
+                               Luv2ShopValidators.notOnlyWhitespace]),
+                               
+        email: new FormControl('',
+                              [Validators.required, Validators.pattern('^[a-z0-9._%+-]+@[a-z0-9.-]+\\.[a-z]{2,4}$')])
       }),
       shippingAddress: this.formBuilder.group({
-        country: new FormControl('', [Validators.required]),
-        street: new FormControl('', [Validators.required, Validators.minLength(2), Luv2ShopValidators.notOnlyWhitespace]),
-        city: new FormControl('', [Validators.required, Validators.minLength(2), Luv2ShopValidators.notOnlyWhitespace]),
+        street: new FormControl('', [Validators.required, Validators.minLength(2), 
+                                     Luv2ShopValidators.notOnlyWhitespace]),
+        city: new FormControl('', [Validators.required, Validators.minLength(2), 
+                                   Luv2ShopValidators.notOnlyWhitespace]),
         state: new FormControl('', [Validators.required]),
-        zipCode: new FormControl('', [Validators.required, Validators.minLength(2), Luv2ShopValidators.notOnlyWhitespace])
+        country: new FormControl('', [Validators.required]),
+        zipCode: new FormControl('', [Validators.required, Validators.minLength(2), 
+                                      Luv2ShopValidators.notOnlyWhitespace])
       }),
       billingAddress: this.formBuilder.group({
-        country: new FormControl('', [Validators.required]),
-        street: new FormControl('', [Validators.required, Validators.minLength(2), Luv2ShopValidators.notOnlyWhitespace]),
-        city: new FormControl('', [Validators.required, Validators.minLength(2), Luv2ShopValidators.notOnlyWhitespace]),
+        street: new FormControl('', [Validators.required, Validators.minLength(2), 
+                                     Luv2ShopValidators.notOnlyWhitespace]),
+        city: new FormControl('', [Validators.required, Validators.minLength(2), 
+                                   Luv2ShopValidators.notOnlyWhitespace]),
         state: new FormControl('', [Validators.required]),
-        zipCode: new FormControl('', [Validators.required, Validators.minLength(2), Luv2ShopValidators.notOnlyWhitespace])
+        country: new FormControl('', [Validators.required]),
+        zipCode: new FormControl('', [Validators.required, Validators.minLength(2), 
+                                      Luv2ShopValidators.notOnlyWhitespace])
       }),
       creditCard: this.formBuilder.group({
         cardType: new FormControl('', [Validators.required]),
-        nameOnCard: new FormControl('', [Validators.required, Validators.minLength(2), Luv2ShopValidators.notOnlyWhitespace]),
+        nameOnCard:  new FormControl('', [Validators.required, Validators.minLength(2), 
+                                          Luv2ShopValidators.notOnlyWhitespace]),
         cardNumber: new FormControl('', [Validators.required, Validators.pattern('[0-9]{16}')]),
         securityCode: new FormControl('', [Validators.required, Validators.pattern('[0-9]{3}')]),
         expirationMonth: [''],
         expirationYear: ['']
-      }),
-
+      })
     });
 
-    //populate credit card months
+    // populate credit card months
+
     const startMonth: number = new Date().getMonth() + 1;
     console.log("startMonth: " + startMonth);
 
@@ -83,50 +99,55 @@ export class CheckoutComponent implements OnInit {
         this.creditCardMonths = data;
       }
     );
-    //populate credit card years
+
+    // populate credit card years
+
     this.luv2ShopFormService.getCreditCardYears().subscribe(
       data => {
         console.log("Retrieved credit card years: " + JSON.stringify(data));
         this.creditCardYears = data;
       }
     );
-    //populate countries
+
+    // populate countries
+
     this.luv2ShopFormService.getCountries().subscribe(
       data => {
         console.log("Retrieved countries: " + JSON.stringify(data));
         this.countries = data;
       }
     );
-
-
-
   }
 
   reviewCartDetails() {
+
+    // subscribe to cartService.totalQuantity
     this.cartService.totalQuantity.subscribe(
       totalQuantity => this.totalQuantity = totalQuantity
     );
-    
+
+    // subscribe to cartService.totalPrice
     this.cartService.totalPrice.subscribe(
       totalPrice => this.totalPrice = totalPrice
     );
+
   }
 
   get firstName() { return this.checkoutFormGroup.get('customer.firstName'); }
   get lastName() { return this.checkoutFormGroup.get('customer.lastName'); }
   get email() { return this.checkoutFormGroup.get('customer.email'); }
 
-  get shippingAddressCountry() { return this.checkoutFormGroup.get('shippingAddress.country'); }
   get shippingAddressStreet() { return this.checkoutFormGroup.get('shippingAddress.street'); }
   get shippingAddressCity() { return this.checkoutFormGroup.get('shippingAddress.city'); }
   get shippingAddressState() { return this.checkoutFormGroup.get('shippingAddress.state'); }
   get shippingAddressZipCode() { return this.checkoutFormGroup.get('shippingAddress.zipCode'); }
+  get shippingAddressCountry() { return this.checkoutFormGroup.get('shippingAddress.country'); }
 
-  get billingAddressCountry() { return this.checkoutFormGroup.get('billingAddress.country'); }
   get billingAddressStreet() { return this.checkoutFormGroup.get('billingAddress.street'); }
   get billingAddressCity() { return this.checkoutFormGroup.get('billingAddress.city'); }
   get billingAddressState() { return this.checkoutFormGroup.get('billingAddress.state'); }
   get billingAddressZipCode() { return this.checkoutFormGroup.get('billingAddress.zipCode'); }
+  get billingAddressCountry() { return this.checkoutFormGroup.get('billingAddress.country'); }
 
   get creditCardType() { return this.checkoutFormGroup.get('creditCard.cardType'); }
   get creditCardNameOnCard() { return this.checkoutFormGroup.get('creditCard.nameOnCard'); }
@@ -135,139 +156,120 @@ export class CheckoutComponent implements OnInit {
 
 
 
-  getStates(formGroupName: string) {
+  copyShippingAddressToBillingAddress(checked: boolean) {
 
-    const formGroup = this.checkoutFormGroup.get(formGroupName);
+    if (checked) {
+      this.checkoutFormGroup.controls?.['billingAddress']
+            .setValue(this.checkoutFormGroup.controls?.['shippingAddress'].value);
 
-    const countryCode = formGroup!.value.country.code;
-    const countryName = formGroup!.value.country.name;
+      // bug fix for states
+      this.billingAddressStates = this.shippingAddressStates;
 
-    console.log(`${formGroupName} country code: ${countryCode}`);
-    console.log(`${formGroupName} country name: ${countryName}`);
+    }
+    else {
+      this.checkoutFormGroup.controls?.['billingAddress'].reset();
 
-    this.luv2ShopFormService.getStates(countryCode).subscribe(
-      data => {
-
-        if (formGroupName === 'shippingAddress') {
-          this.shippingAddressStates = data;
-        }
-        else {
-          this.billingAddressStates = data;
-        }
-
-        // select first item by default
-        formGroup!.get('state')!.setValue(data[0]);
-      }
-    );
+      // bug fix for states
+      this.billingAddressStates = [];
+    }
+    
   }
 
   onSubmit() {
-    
-    if(this.checkoutFormGroup.invalid){
+    console.log("Handling the submit button");
+
+    if (this.checkoutFormGroup.invalid) {
       this.checkoutFormGroup.markAllAsTouched();
       return;
     }
-    console.log("CheckoutFormGroup is valid: " + this.checkoutFormGroup.valid);
 
-    //set up order
-    let order = new Order();
-    order.totalPrice = this.totalPrice;
-    order.totalQuantity = this.totalQuantity;
+    // set up order
+    let order = new Order(this.totalQuantity, this.totalPrice);
 
-    //get cart items
+    // get cart items
     const cartItems = this.cartService.cartItems;
 
-    //create orderItems from cartItems
-    let orderItems: OrderItem[] = cartItems.map(tempCartItem => new OrderItem(tempCartItem));
-
-    /*same as :
-    let orderItems: OrderItems[] = [];
-    for(let i=0; i < cartItems.length; i++){
+    // create orderItems from cartItems
+    // - long way
+    /*
+    let orderItems: OrderItem[] = [];
+    for (let i=0; i < cartItems.length; i++) {
       orderItems[i] = new OrderItem(cartItems[i]);
     }
     */
 
-    //set up purchase
+    // - short way of doing the same thingy
+    let orderItems: OrderItem[] = cartItems.map(tempCartItem => new OrderItem(tempCartItem.imageUrl!, tempCartItem.unitPrice!, tempCartItem.quantity, tempCartItem.id!));
+
+    // set up purchase
     let purchase = new Purchase();
-
-    //populate purchase - customer
+    
+    // populate purchase - customer
     purchase.customer = this.checkoutFormGroup.controls['customer'].value;
-
-
-    //populate purchase - shipping address
+    
+    // populate purchase - shipping address
     purchase.shippingAddress = this.checkoutFormGroup.controls['shippingAddress'].value;
     const shippingState: State = JSON.parse(JSON.stringify(purchase.shippingAddress.state));
     const shippingCountry: Country = JSON.parse(JSON.stringify(purchase.shippingAddress.country));
     purchase.shippingAddress.state = shippingState.name;
     purchase.shippingAddress.country = shippingCountry.name;
 
-    //populate purchase - billing address
+    // populate purchase - billing address
     purchase.billingAddress = this.checkoutFormGroup.controls['billingAddress'].value;
     const billingState: State = JSON.parse(JSON.stringify(purchase.billingAddress.state));
     const billingCountry: Country = JSON.parse(JSON.stringify(purchase.billingAddress.country));
     purchase.billingAddress.state = billingState.name;
     purchase.billingAddress.country = billingCountry.name;
-
-    //populate purchase - order and orderItems
+  
+    // populate purchase - order and orderItems
     purchase.order = order;
     purchase.orderItems = orderItems;
 
-    //call REST API via the checkoutService
+    // call REST API via the CheckoutService
     this.checkoutService.placeOrder(purchase).subscribe({
         next: response => {
-          alert(`Your order has been recieved.\nOrder tracking number: ${response.orderTrackingNumber}`);
-          //reset cart
+          alert(`Your order has been received.\nOrder tracking number: ${response.orderTrackingNumber}`);
+
+          // reset cart
           this.resetCart();
+
         },
         error: err => {
           alert(`There was an error: ${err.message}`);
         }
       }
     );
+
   }
-  resetCart(){
-    //reset cart data
+
+  resetCart() {
+    // reset cart data
     this.cartService.cartItems = [];
     this.cartService.totalPrice.next(0);
     this.cartService.totalQuantity.next(0);
-
-    //reset the form
+    
+    // reset the form
     this.checkoutFormGroup.reset();
 
-    //navigate back to the products page
+    // navigate back to the products page
     this.router.navigateByUrl("/products");
   }
 
-  copyShippingAddrToBillingAddr(event: { target: any; }) {
-
-    if (event.target!.checked) {
-      this.checkoutFormGroup.controls['billingAddress'].setValue(this.checkoutFormGroup.controls['shippingAddress'].value);
-
-      //bug fix for states
-      this.billingAddressStates = this.shippingAddressStates;
-
-    } else {
-      this.checkoutFormGroup.controls['billingAddress'].reset();
-
-      // bug fix for states
-      this.billingAddressStates = [];
-    }
-
-  }
-
   handleMonthsAndYears() {
+
     const creditCardFormGroup = this.checkoutFormGroup.get('creditCard');
 
     const currentYear: number = new Date().getFullYear();
     const selectedYear: number = Number(creditCardFormGroup?.value.expirationYear);
 
-    //if the current year equals the selected year, then start with current month
+    // if the current year equals the selected year, then start with the current month
 
     let startMonth: number;
 
     if (currentYear === selectedYear) {
       startMonth = new Date().getMonth() + 1;
-    } else {
+    }
+    else {
       startMonth = 1;
     }
 
@@ -279,4 +281,29 @@ export class CheckoutComponent implements OnInit {
     );
   }
 
+  getStates(formGroupName: string) {
+
+    const formGroup = this.checkoutFormGroup.get(formGroupName);
+
+    const countryCode = formGroup?.value.country.code;
+    const countryName = formGroup?.value.country.name;
+
+    console.log(`${formGroupName} country code: ${countryCode}`);
+    console.log(`${formGroupName} country name: ${countryName}`);
+
+    this.luv2ShopFormService.getStates(countryCode).subscribe(
+      data => {
+
+        if (formGroupName === 'shippingAddress') {
+          this.shippingAddressStates = data; 
+        }
+        else {
+          this.billingAddressStates = data;
+        }
+
+        // select first item by default
+        formGroup?.get('state')?.setValue(data[0]);
+      }
+    );
+  }
 }
