@@ -1,44 +1,41 @@
 import { Component, Inject, OnInit } from '@angular/core';
-import { OKTA_AUTH, OktaAuthStateService } from '@okta/okta-angular';
-import OktaAuth from '@okta/okta-auth-js';
+ import {OktaAuthStateService, OKTA_AUTH} from '@okta/okta-angular';
+ import {OktaAuth} from '@okta/okta-auth-js'
 
-@Component({
-  selector: 'app-login-status',
-  templateUrl: './login-status.component.html',
-  styleUrls: ['./login-status.component.css']
-})
-export class LoginStatusComponent implements OnInit{
+ @Component({
+   selector: 'app-login-status',
+   templateUrl: './login-status.component.html',
+   styleUrls: ['./login-status.component.css']
+ })
+ export class LoginStatusComponent implements OnInit{
+   isAuthenticated:boolean = false;
+   userFullName: string = '';
 
-  isAuthenticated: boolean = false;
-  userFullName: string = '';
+   constructor(private oktaAuthService: OktaAuthStateService,
+             @Inject(OKTA_AUTH) private oktaAuth: OktaAuth){}
+   ngOnInit(): void {
 
-  constructor(private oktaAuthService: OktaAuthStateService,
-               @Inject(OKTA_AUTH) private oktaAuth: OktaAuth){}
+     //Subscribe to authentication state changes
 
-  ngOnInit(): void {
-    //subscribe to authentication state changes
-    this.oktaAuthService.authState$.subscribe(
-      (result) => {
-        this.isAuthenticated = result.isAuthenticated!;
-        this.getUserDetails();
-      }
-    );
-  }
+     this.oktaAuthService.authState$.subscribe(
+       (result) => {
+         this.isAuthenticated = result.isAuthenticated!;
+         this.getUserDetails();
+       }
+     )
+   }
+   getUserDetails() {
+     // Fetch the logged-in user details (user's claims)
+     // User full name is exposed as a property named 'name'
+     this.oktaAuth.getUser().then(
+       (res) => {
+         this.userFullName = res.name as string;
+       }
+     );
+   }
 
-  getUserDetails() {
-    if(this.isAuthenticated){
-      //Fetch the logged in user details(users claims)
-      //user full name is exposed as a property name
-      this.oktaAuth.getUser().then(
-        (res) => { this.userFullName = res.name as string; }  
-      );
-
-    }
-  }
-
-  logout(){
-    //Terminate the session with Okta removes current tokens.
-    this.oktaAuth.signOut();
-  }
-
-}
+   logout(){
+     //Terminated the session with Okta and removes current tokens.
+     this.oktaAuth.signOut();
+   }
+ }
