@@ -1,46 +1,50 @@
 import { Component, Inject, OnInit } from '@angular/core';
 import { OKTA_AUTH } from '@okta/okta-angular';
-import OktaAuth from '@okta/okta-auth-js';
-import OktaSignIn from '@okta/okta-signin-widget';
+import { OktaAuth } from '@okta/okta-auth-js';
 import myAppConfig from 'src/app/config/my-app-config';
+import OktaSignIn from '@okta/okta-signin-widget';
 
 @Component({
   selector: 'app-login',
   templateUrl: './login.component.html',
   styleUrls: ['./login.component.css']
 })
-export class LoginComponent implements OnInit{
+export class LoginComponent implements OnInit {
 
   oktaSignin: any;
 
-  constructor(@Inject(OKTA_AUTH) private oktaAuth: OktaAuth){
+  constructor(@Inject(OKTA_AUTH) public oktaAuth: OktaAuth) {
     this.oktaSignin = new OktaSignIn({
       logo: 'assets/images/online-shop-logo.jpg',
-      baseUrl: myAppConfig.oidc.issuer.split('/oauth2')[0],
+      features: {
+        registration: true
+      },
+      baseUrl: myAppConfig.oidc.issuer.split('/outh2')[0],
       clientId: myAppConfig.oidc.clientId,
       redirectUri: myAppConfig.oidc.redirectUri,
+      useClassicEngine: true,
       authParams: {
         pkce: true,
         issuer: myAppConfig.oidc.issuer,
         scopes: myAppConfig.oidc.scopes
       }
     });
-  }
+   }
 
   ngOnInit(): void {
     this.oktaSignin.remove();
 
     this.oktaSignin.renderEl({
-      el: '#okta-sign-in-widget'},//same name as div tag id in login
+      el: '#okta-sign-in-widget'
+    }, // this name should be same as div tag in the login.component.html
       (response: any) => {
-        if(response.status === 'SUCCESS'){
+        if(response.status === 'SUCCESS') {
           this.oktaAuth.signInWithRedirect();
         }
       },
       (error: any) => {
         throw error;
       }
-      );
+    );
   }
-
 }
