@@ -1,4 +1,4 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ProductCategoryMenuComponent } from './product-category-menu.component';
 
@@ -6,10 +6,14 @@ describe('ProductCategoryMenuComponent', () => {
   let component: ProductCategoryMenuComponent;
   let fixture: ComponentFixture<ProductCategoryMenuComponent>;
 
-  beforeEach(() => {
+  beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ProductCategoryMenuComponent]
-    });
+      declarations: [ ProductCategoryMenuComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
     fixture = TestBed.createComponent(ProductCategoryMenuComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
