@@ -10,22 +10,21 @@ import { ProductService } from 'src/app/services/product.service';
 export class ProductCategoryMenuComponent implements OnInit {
 
   productCategories: ProductCategory[] = [];
+  
+  constructor(private productService: ProductService) { }
 
-  constructor(private productService: ProductService){}
-
-  ngOnInit(): void {
+  ngOnInit() {
     this.listProductCategories();
   }
 
   listProductCategories() {
+
     this.productService.getProductCategories().subscribe(
       data => {
         console.log('Product Categories=' + JSON.stringify(data));
         this.productCategories = data;
       }
-    )
+    );
   }
-
-
 
 }
