@@ -1,19 +1,22 @@
 import { Component, OnInit } from '@angular/core';
-import { ActivatedRoute } from '@angular/router';
-import { CartItem } from 'src/app/common/cart-item';
 import { Product } from 'src/app/common/product';
-import { CartService } from 'src/app/services/cart-service';
 import { ProductService } from 'src/app/services/product.service';
+import { ActivatedRoute } from '@angular/router';
+import { CartService } from 'src/app/services/cart.service';
+import { CartItem } from 'src/app/common/cart-item';
 
 @Component({
   selector: 'app-product-details',
   templateUrl: './product-details.component.html',
   styleUrls: ['./product-details.component.css']
 })
-export class ProductDetailsComponent implements OnInit{
- product!: Product;
+export class ProductDetailsComponent implements OnInit {
 
- constructor(private productService: ProductService, private route: ActivatedRoute, private cartService: CartService){}
+  product: Product = new Product();
+
+  constructor(private productService: ProductService,
+              private cartService: CartService,
+              private route: ActivatedRoute) { }
 
   ngOnInit(): void {
     this.route.paramMap.subscribe(() => {
@@ -22,7 +25,8 @@ export class ProductDetailsComponent implements OnInit{
   }
 
   handleProductDetails() {
-    //get the 'id' param string and convert it to a number
+
+    // get the "id" param string. convert string to a number using the "+" symbol
     const theProductId: number = +this.route.snapshot.paramMap.get('id')!;
 
     this.productService.getProduct(theProductId).subscribe(
@@ -32,11 +36,13 @@ export class ProductDetailsComponent implements OnInit{
     )
   }
 
-  addToCart(){
+  addToCart() {
+
     console.log(`Adding to cart: ${this.product.name}, ${this.product.unitPrice}`);
-
-    const theCartItem = new CartItem(this.product);
-
+    let theCartItem = new CartItem(this.product.id, this.product.name, this.product.imageUrl, this.product.unitPrice);
+    
     this.cartService.addToCart(theCartItem);
+    
   }
+
 }
