@@ -1,9 +1,9 @@
 import { Component, OnInit } from '@angular/core';
+import { ProductService } from 'src/app/services/product.service';
+import { Product } from 'src/app/common/product';
 import { ActivatedRoute } from '@angular/router';
 import { CartItem } from 'src/app/common/cart-item';
-import { Product } from 'src/app/common/product';
-import { ProductService } from 'src/app/services/product.service';
-import { CartService } from 'src/app/services/cart-service';
+import { CartService } from 'src/app/services/cart.service';
 
 @Component({
   selector: 'app-product-list',
@@ -16,19 +16,22 @@ export class ProductListComponent implements OnInit {
   currentCategoryId: number = 1;
   previousCategoryId: number = 1;
   searchMode: boolean = false;
-  //properties for pagination
+
+  // new properties for pagination
   thePageNumber: number = 1;
   thePageSize: number = 5;
   theTotalElements: number = 0;
 
   previousKeyword: string = "";
-  
 
+  constructor(private productService: ProductService,
+              private cartService: CartService,
+              private route: ActivatedRoute) { }
 
-  constructor(private productService: ProductService, private route: ActivatedRoute, private cartService: CartService) { }
-
-  ngOnInit(): void {
-    this.route.paramMap.subscribe(() => { this.listProducts() });
+  ngOnInit() {
+    this.route.paramMap.subscribe(() => {
+      this.listProducts();
+    });
   }
 
   listProducts() {
@@ -37,15 +40,21 @@ export class ProductListComponent implements OnInit {
 
     if (this.searchMode) {
       this.handleSearchProducts();
-    } else {
+    }
+    else {
       this.handleListProducts();
     }
+
   }
 
   handleSearchProducts() {
-    const theKeyword: string = this.route.snapshot.paramMap.get(`keyword`)!;
 
-    if(this.previousKeyword != theKeyword){
+    const theKeyword: string = this.route.snapshot.paramMap.get('keyword')!;
+
+    // if we have a different keyword than previous
+    // then set thePageNumber to 1
+
+    if (this.previousKeyword != theKeyword) {
       this.thePageNumber = 1;
     }
 
@@ -53,26 +62,35 @@ export class ProductListComponent implements OnInit {
 
     console.log(`keyword=${theKeyword}, thePageNumber=${this.thePageNumber}`);
 
-    this.productService.SearchProductPaginate(this.thePageNumber - 1, this.thePageSize, theKeyword).subscribe(
-      this.processResult()
-    );
+    // now search for the products using keyword
+    this.productService.searchProductsPaginate(this.thePageNumber - 1,
+                                               this.thePageSize,
+                                               theKeyword).subscribe(this.processResult());
+                                               
   }
 
   handleListProducts() {
-    //check if 'id' is available
+
+    // check if "id" parameter is available
     const hasCategoryId: boolean = this.route.snapshot.paramMap.has('id');
 
     if (hasCategoryId) {
-      //get the 'id' param string. convert string to a number using the "+" symbol
+      // get the "id" param string. convert string to a number using the "+" symbol
       this.currentCategoryId = +this.route.snapshot.paramMap.get('id')!;
-    } else {
-      //not category id available ... default to category id 1
+    }
+    else {
+      // not category id available ... default to category id 1
       this.currentCategoryId = 1;
     }
 
-    //check if we have a different category than previous
-    //Note: Angular will reuse a component if its currently being viewed
-    if(this.previousCategoryId != this.currentCategoryId){
+    //
+    // Check if we have a different category than previous
+    // Note: Angular will reuse a component if it is currently being viewed
+    //
+
+    // if we have a different category id than previous
+    // then set thePageNumber back to 1
+    if (this.previousCategoryId != this.currentCategoryId) {
       this.thePageNumber = 1;
     }
 
@@ -80,31 +98,35 @@ export class ProductListComponent implements OnInit {
 
     console.log(`currentCategoryId=${this.currentCategoryId}, thePageNumber=${this.thePageNumber}`);
 
-    //now get the product for the given category id
-    this.productService.getProductListPaginate(this.thePageNumber - 1, this.thePageSize, this.currentCategoryId)
-                       .subscribe(this.processResult());
+    // now get the products for the given category id
+    this.productService.getProductListPaginate(this.thePageNumber - 1,
+                                               this.thePageSize,
+                                               this.currentCategoryId)
+                                               .subscribe(this.processResult());
   }
 
-  updatePageSize(pageSize: string){
+  updatePageSize(pageSize: string) {
     this.thePageSize = +pageSize;
     this.thePageNumber = 1;
     this.listProducts();
   }
 
-  private processResult(){
-    return (data: any) => {this.products = data._embedded.products;
+  processResult() {
+    return (data: any) => {
+      this.products = data._embedded.products;
       this.thePageNumber = data.page.number + 1;
       this.thePageSize = data.page.size;
       this.theTotalElements = data.page.totalElements;
     };
   }
 
-  addToCart(theProduct: Product){
+  addToCart(theProduct: Product) {
+    
     console.log(`Adding to cart: ${theProduct.name}, ${theProduct.unitPrice}`);
 
-    const theCartItem = new CartItem(theProduct);
-    
-    //Todo ... do the real work
+    // TODO ... do the real work
+    let theCartItem = new CartItem(theProduct.id, theProduct.name, theProduct.imageUrl, theProduct.unitPrice);
+
     this.cartService.addToCart(theCartItem);
   }
 
