@@ -1,9 +1,10 @@
 export default {
 
-    oidc:{
-        clientId: '0oaeley9cv9Ie4NFg5d7',
-        issuer: 'https://dev-07309204.okta.com/oauth2/default',
+    oidc: {
+        clientId: '0oaeokues58MIt7uq5d7',
+        issuer: 'https://dev-00391550.okta.com/oauth2/default',
         redirectUri: 'http://localhost:4200/login/callback',
-        scopes: ['openid', 'profile', 'email']
+        scopes:['openid','profile','email']
     }
+
 }
