@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import { CartItem } from '../common/cart-item';
-import { BehaviorSubject, ReplaySubject, Subject } from 'rxjs';
+import { BehaviorSubject, Subject } from 'rxjs';
+import { Product } from '../common/product';
 
 @Injectable({
   providedIn: 'root'
@@ -14,76 +15,88 @@ export class CartService {
 
   constructor() { }
 
-  addToCart(theCartItem: CartItem){
-    //check if we already have the item in our cart
+  addToCart(theCartItem: CartItem) {
+
+    // check if we already have the item in our cart
     let alreadyExistsInCart: boolean = false;
-    let existingCartItem!: CartItem;
+    // let existingCartItem: CartItem = undefined;
+    let existingCartItem: CartItem = new CartItem();
 
-    if(this.cartItems.length > 0){
-      //find the item in the cart base on item id
+    if (this.cartItems.length > 0) {
+      // find the item in the cart based on item id
 
-      existingCartItem = this.cartItems.find(tempCartItem => tempCartItem.id === theCartItem.id)!;
-      //check if we found it
+      existingCartItem = this.cartItems.find( tempCartItem => tempCartItem.id === theCartItem.id )!;
+
+      // check if we found it
       alreadyExistsInCart = (existingCartItem != undefined);
     }
-    if(alreadyExistsInCart){
-      //increment the quantity
+
+    if (alreadyExistsInCart) {
+      // increment the quantity
       existingCartItem.quantity++;
-    }else{
-      //just add the item to the array
+    }
+    else {
+      // just add the item to the array
       this.cartItems.push(theCartItem);
     }
-    //compute cart quantity and cart total
+
+    // compute cart total price and total quantity
     this.computeCartTotals();
   }
 
-  computeCartTotals(){
+  computeCartTotals() {
+
     let totalPriceValue: number = 0;
     let totalQuantityValue: number = 0;
 
-    for(let currentCartItem of this.cartItems){
-      totalPriceValue += currentCartItem.quantity * currentCartItem.unitPrice;
+    for (let currentCartItem of this.cartItems) {
+      totalPriceValue += currentCartItem.quantity * currentCartItem.unitPrice!;
       totalQuantityValue += currentCartItem.quantity;
     }
 
-    //publish the new values
+    // publish the new values ... all subscribers will receive the new data
     this.totalPrice.next(totalPriceValue);
     this.totalQuantity.next(totalQuantityValue);
 
-    //log cart data just for debugging purposes
+    // log cart data just for debugging purposes
     this.logCartData(totalPriceValue, totalQuantityValue);
   }
 
   logCartData(totalPriceValue: number, totalQuantityValue: number) {
-    console.log('Content of the cart');
-    for(let tempCartItem of this.cartItems){
-      const subTotalPrice = tempCartItem.quantity * tempCartItem.unitPrice;
+
+    console.log('Contents of the cart');
+    for (let tempCartItem of this.cartItems) {
+      const subTotalPrice = tempCartItem.quantity * tempCartItem.unitPrice!;
       console.log(`name: ${tempCartItem.name}, quantity=${tempCartItem.quantity}, unitPrice=${tempCartItem.unitPrice}, subTotalPrice=${subTotalPrice}`);
     }
 
     console.log(`totalPrice: ${totalPriceValue.toFixed(2)}, totalQuantity: ${totalQuantityValue}`);
-    console.log(`-----------`);
+    console.log('----');
   }
 
   decrementQuantity(theCartItem: CartItem) {
+
     theCartItem.quantity--;
-    if(theCartItem.quantity === 0){
+
+    if (theCartItem.quantity === 0) {
       this.remove(theCartItem);
-    }else{
+    }
+    else {
       this.computeCartTotals();
     }
   }
-  remove(theCartItem: CartItem) {
-    //get index of item in the array
-    const itemIndex = this.cartItems.findIndex(tempCartItem => tempCartItem.id == theCartItem.id);
 
-    //if found, remove the item from the array at the given index
-    if(itemIndex > -1){
+  remove(theCartItem: CartItem) {
+
+    // get index of item in the array
+    const itemIndex = this.cartItems.findIndex( tempCartItem => tempCartItem.id === theCartItem.id );
+
+    // if found, remove the item from the array at the given index
+    if (itemIndex > -1) {
       this.cartItems.splice(itemIndex, 1);
 
       this.computeCartTotals();
     }
   }
 
-  
 }
