@@ -1,7 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Purchase } from '../common/purchase';
 import { Observable } from 'rxjs';
+import { Purchase } from '../common/purchase';
 
 @Injectable({
   providedIn: 'root'
@@ -12,7 +12,8 @@ export class CheckoutService {
 
   constructor(private httpClient: HttpClient) { }
 
-  placeOrder(purchase: Purchase): Observable<any>{
-    return this.httpClient.post<Purchase>(this.purchaseUrl, purchase);
+  placeOrder(purchase: Purchase): Observable<any> {
+    return this.httpClient.post<Purchase>(this.purchaseUrl, purchase);    
   }
+  
 }
